requirements.txt
fastapi
uvicorn
pandas
joblib
pytz
google-cloud-bigquery
pandas-gbq
scikit-learn